+++ lib/implicit_navigator.dart
@@ -1,5 +1,6 @@
 export 'src/implicit_navigator.dart'
     show
+        ImplicitLocalHistoryEntry,
         ImplicitNavigator,
         ImplicitNavigatorBackButton,
         ImplicitNavigatorState,